Реализовать механизм дополнительных работ для позиций заказчика
Добавлена полная поддержка создания и редактирования дополнительных работ (ДОП) для позиций заказчика.

Основные изменения:

1. Создан компонент AddAdditionalPositionModal.tsx:
   - Модальное окно для добавления дополнительных работ
   - Поля: наименование работы, единица измерения, количество ГП, примечание ГП
   - Автоматический расчет номера позиции с суффиксом (5.1, 5.2, 5.3)
   - Загрузка единиц измерения из справочника units

2. Доработан ClientPositions.tsx:
   - Добавлена кнопка "+" в колонке "Итого" для создания доп работ
   - Реализовано отображение метки "ДОП" вместо номера раздела
   - Добавлен визуальный отступ (20px) только для доп работ
   - Объединены две карточки (фильтры и данные тендера) в одну
   - Увеличены фильтры на 30% (span с 7 до 9)
   - Убран прочерк "-" в колонке "Итого" для неконечных позиций
   - Запрещено создание доп работ у доп работ (кнопка скрыта)

3. Доработан PositionItems.tsx:
   - Добавлена метка "ДОП" для дополнительных работ
   - Реализовано редактирование наименования и единицы измерения для доп работ
   - Убраны итоговые теги ("Итого: ..." и "Р X М Y") из шапки позиции
   - Автосохранение данных при изменении

4. База данных:
   - Изменен тип position_number с integer на numeric(10,2) для поддержки decimal номеров
   - Создана миграция 018_change_position_number_to_numeric.sql
   - Исправлена миграция 017 для добавления проверок на существование constraints
   - Обновлена схема prod.sql

Техническая реализация:
- Дополнительные работы привязываются через parent_position_id
- Флаг is_additional = true
- hierarchy_level = parent.hierarchy_level + 1
- position_number вычисляется автоматически (5.1, 5.2, 5.3...)
- Дополнительные работы являются конечными позициями (можно добавлять работы/материалы)

+++ supabase/schemas/prod.sql
@@ -420,7 +420,7 @@ COMMENT ON COLUMN public.boq_items.parent_work_item_id IS 'Привязка ма
 CREATE TABLE IF NOT EXISTS public.client_positions (
     id uuid NOT NULL DEFAULT uuid_generate_v4(),
     tender_id uuid NOT NULL,
-    position_number integer(32) NOT NULL,
+    position_number numeric(10,2) NOT NULL,
     unit_code text,
     volume numeric(18,6),
     client_note text,
